+++ app/src/main/java/com/example/dominika/appobchod/Interview.java
@@ -1,6 +1,7 @@
 package com.example.dominika.appobchod;
 
 import android.content.Intent;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
@@ -17,39 +18,26 @@ public class Interview extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_interview);
 
-        //listenClicks();
+        Button next = (Button) findViewById(R.id.button_add_alergy);
+        assert next != null;
+        next.setOnClickListener(new View.OnClickListener() {
+            public void onClick(View view) {
+                Intent intent = new Intent(Interview.this, ShowAllerginList.class);
+                startActivity(intent);
+            }
+        });
     }
-/*
-    private void listenClicks() {
-        Button add_alergy = (Button) findViewById(R.id.button_add_alergy);
-        Button add_med = (Button) findViewById(R.id.button_add_med);
-        Button add_disease = (Button) findViewById(R.id.button_add_disease);
-        //assert add_alergy != null;
-        assert add_med != null;
-        assert add_disease != null;
 
-        add_alergy.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                ShowAllerginScreen(view);
-            }
-        });
-
-        add_med.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                setContentView(R.layout.activity_alergy);
-            }
-        });
-
-        add_disease.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                setContentView(R.layout.activity_alergy);
-            }
-        });
-    }*/
-
-    public void ShowAllerginScreenLOL(View w){
+    public void ShowAllerginScreen(View w){
         Intent intent = new Intent(Interview.this, ShowAllerginList.class);
         startActivity(intent);
+    }
+
+    public void showAlert(View v) {
+        AlertDialog.Builder ad = new AlertDialog.Builder(this);
+        ad.setMessage("Gratulacje lekarzu! Dodano pacjenta na oddział!").create();
+        ad.setTitle("Komunikat");
+        ad.show();
     }
 
     //TODO walidacja pesel
