+++ app/src/main/res/layout/activity_first_screen.xml
@@ -19,73 +19,33 @@
 
     <Button
         android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="80dp"
+        android:layout_marginTop="40dp"
         android:layout_below="@+id/textView"
         android:layout_alignParentRight="true"
         android:layout_alignParentEnd="true"
-        android:layout_marginTop="35dp"
         android:id="@+id/doctor"
         android:text="Lekarz"
-       android:onClick="showLogin"/>
+        android:onClick="DoctorButtonOnClick"/>
 
     <Button
         android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="80dp"
+        android:layout_marginTop="40dp"
         android:id="@+id/nurse"
         android:text="Pielęgniarka"
         android:layout_below="@+id/doctor"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:onClick="DoctorButtonOnClick" />
+        android:onClick="NurseButtonOnClick" />
 
     <Button
         android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="80dp"
+        android:layout_marginTop="40dp"
         android:id="@+id/tech_man"
         android:text="Zaopatrzenie"
         android:layout_below="@+id/nurse"
         android:layout_centerHorizontal="true"
         android:onClick="TechnicalButtonOnClick" />
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="646dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="85dp"
-        android:visibility="invisible"
-        android:layout_below="@+id/tech_man"
-        android:layout_centerHorizontal="true">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Login:"
-            android:textSize="20dp"/>
-
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Hasło:"
-            android:textSize="20dp"/>
-
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-        <Button
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:text="Zapisz"
-            android:background="#4CAF50"
-            android:layout_marginTop="30dp"
-            android:onClick="showAlert"/>
-    </LinearLayout>
-
-
-
 </RelativeLayout>
