+++ app/src/main/res/layout/activity_interview.xml
@@ -204,9 +204,9 @@
                         android:layout_width="38dp"
                         android:layout_height="38dp"
                         android:text="+"
-                        android:background="#4CAF50"
+                        android:background="@color/buttonColor"
                         android:id="@+id/button_add_alergy"
-                        android:onClick="ShowAllerginScreenLOL" />
+                        android:onClick="ShowAllerginScreen" />
 
                     <TextView
                         android:layout_width="wrap_content"
@@ -234,7 +234,7 @@
                         android:layout_width="38dp"
                         android:layout_height="38dp"
                         android:text="+"
-                        android:background="#4CAF50"
+                        android:background="@color/buttonColor"
                         android:id="@+id/button_add_med" />
 
                     <TextView
@@ -263,7 +263,7 @@
                         android:layout_width="38dp"
                         android:layout_height="38dp"
                         android:text="+"
-                        android:background="#4CAF50"
+                        android:background="@color/buttonColor"
                         android:id="@+id/button_add_disease" />
 
                     <TextView
@@ -300,7 +300,7 @@
                     android:layout_width="fill_parent"
                     android:layout_height="wrap_content"
                     android:text="Zapisz"
-                    android:background="#4CAF50"
+                    android:background="@color/buttonColor"
                     android:layout_marginTop="30dp"
                     android:onClick="showAlert"/>
